Added officer functionality -- we can now have managers log in.  They have access to pages with require_users that match their names.  Huge, huge mess.  Here's a list of changes: set_passwd now takes fourth optional argument, .  If true, it calls check_officer_passwd() instead of check_passwd(), and replaces the password in officer_password_table instead of password_table.  check_officer_passwd() has been added as a function.  It does a bit less error-checking than check_passwd().  add_authorized_user() will use officer_name if it needs to when logging (it prefers to use member_name, since that's more informative for readers of the log).  In an unrelated change, it no longer logs the full list of privileges, since it's only ever called with one privilege change at a time, so just that privilege change is logged.  get_session_member() is now called with an optional  argument, which will make it look at ['officer_session_id'] instead of just 'session_id'.  set_session() now takes two additional optional arguments,  and .   does the usual.   allows us to delete the session whose id is given by member_name and not make a new one.  set_session also now clears any old cookies before it sets the new one.
require_user() has had a lot of changes.  It now sets the global variable .  It makes  into an array when it can -- the default way to specify that you want to require a user should be an empty array now.  It no longer has the starred options for privileges -- shouldn't be necessary with officers.  Speaking of officers, it now checks for a valid officer.  If the officer matches a privilege, require_user basically goes into ok_nouser mode -- it does what it can, and then returns true.  If the officer can't be authenticated, it does what it can, and then goes to member_check.  Finally, if the officer's password isn't set, it goes straight to set_passwd.  The logic for normal authentication has been cleaned up a bit.  The logout button now shows all logged-in ids, including a superuser.  By the way, get_user_privileges will now return an empty array instead of an array with one empty element.